+++ CHANGES.rst
@@ -13,6 +13,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- #451: Properly test and reject empty submit requests.
+
 - #376: Document the added home mcc/mnc fields.
 
 - #419: Update geolocate docs to mention all GLS fields.